Trait tidefi_primitives::pallet::StakingExt

pub trait StakingExt<AccountId> {
    fn on_session_end(
        session_index: SessionIndex,
        session_trade_values: Vec<(CurrencyId, Balance)>,
        fees_account_id: AccountId
    ) -> Result<(), DispatchError>;
    fn account_id() -> AccountId;
    fn account_stakes_size() -> u64;
}

Required Methods

fn on_session_end(
session_index: SessionIndex,
session_trade_values: Vec<(CurrencyId, Balance)>,
fees_account_id: AccountId
) -> Result<(), DispatchError>

Triggered when a session end in the Fee pallet

fn account_id() -> AccountId

Get the staking account id where the funds are transfered

fn account_stakes_size() -> u64

Get the size of AccountStakes, used to predict the weight of on_session_end