Crate tidefi_primitives

Modules

assets

networks

pallet

Structs

ActiveEraInfo

Information regarding the active era (era in used in session).

BalanceInfo

CurrencyBalance

Currency balance.

CurrencyMetadata

Currency metadata.

Fee

Information regarding a fee

Stake

Stake details.

StakeCurrencyMeta

Stake currency meta.

Swap

Swap details stored on-chain.

SwapConfirmation

Market maker swap confirmation.

SwapExtrinsic

Swap extrinsic with swap fee details.

Withdrawal

Withdrawal details.

Enums

CurrencyId

Enum indicating the currency. Tide is the native token.

StatusCode

Enum indicating status of the chain.

SwapStatus

Swap status.

WithdrawalStatus

Withdrawal status.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them.

Amount

Signed version of Balance

AssetId

Represent a Wrapped Asset.

Balance

Balance of an account.

Block

Block type.

BlockId

Block ID.

BlockNumber

An index to a block.

DigestItem

Digest item type.

EraIndex

Counter for the number of eras that have passed.

Hash

A hash of some data used by the chain.

Header

Header type.

Index

Index of a transaction in the chain.

Moment

Type used for expressing timestamp.

SessionIndex

Counter for the number of sessions that have passed.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

Timestamp

A timestamp: milliseconds since the unix epoch. u64 is enough to represent a duration of half a billion years, when the time scale is milliseconds.